+++ android/src/main/java/com/visioncameraplugininatvision/VisionCameraPluginInatVisionPlugin.java
@@ -86,11 +86,6 @@ public class VisionCameraPluginInatVisionPlugin extends FrameProcessorPlugin {
       throw new RuntimeException("Taxonomy path is null");
     };
 
-    String patchedOrientationAndroid = (String)arguments.get("patchedOrientationAndroid");
-    if (patchedOrientationAndroid == null) {
-      throw new RuntimeException("patchedOrientationAndroid must be a string passing in the current device orientation");
-    }
-
     // Destructure optional parameters and set values
     String filterByTaxonId = (String)arguments.get("filterByTaxonId");
     setFilterByTaxonId(filterByTaxonId != null ? Integer.valueOf(filterByTaxonId) : null);
@@ -172,7 +167,7 @@ public class VisionCameraPluginInatVisionPlugin extends FrameProcessorPlugin {
     List<Map> cleanedPredictions = new ArrayList<>();
     if (mImageClassifier != null) {
       mImageClassifier.setGeomodelScores(geomodelScores);
-      Bitmap bmp = BitmapUtils.getBitmap(image, patchedOrientationAndroid);
+      Bitmap bmp = BitmapUtils.getBitmap(image);
       Log.d(TAG, "originalBitmap: " + bmp + ": " + bmp.getWidth() + " x " + bmp.getHeight());
       // Crop the center square of the frame
       int minDim = (int) Math.round(Math.min(bmp.getWidth(), bmp.getHeight()) * mCropRatio);